+++ src/config.h
@@ -1,0 +1,10 @@
+#ifndef CONFIG_H
+#define CONFIG_H
+
+// Ensure ArduinoJson uses 64 bit integers to ensure timestamps are stored correctly.
+#define ARDUINOJSON_USE_LONG_LONG 1
+
+
+
+
+#endif