+++ Changelog.md
@@ -1,4 +1,9 @@
 # Changelog
+## [1.0.0] - 27-1-2026
+### Added
+- `ThingDevice`, a class that represents a device in Thingsboard. This class can upload attributes and telemetry to thingsboard.
+- `ThingGateway`, a class that facilitates the communication with a Thingsboard server for a number of `ThingDevices`.
+
 ## [0.0.1] - 26-1-2026
 ### Added
 - gitignore